move: 도메인별 패키지 생성 후 이동
