+++ src/main/java/trendravel/photoravel_be/db/respository/location/LocationRepository.java
@@ -1,4 +1,4 @@
-package trendravel.photoravel_be.db.respository;
+package trendravel.photoravel_be.db.respository.location;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import trendravel.photoravel_be.db.location.Location;
