+++ src/main/java/trendravel/photoravel_be/db/respository/review/ReviewRepository.java
@@ -1,4 +1,4 @@
-package trendravel.photoravel_be.db.respository;
+package trendravel.photoravel_be.db.respository.review;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import trendravel.photoravel_be.db.review.Review;
