+++ src/main/java/trendravel/photoravel_be/db/respository/spot/SpotRepository.java
@@ -1,4 +1,4 @@
-package trendravel.photoravel_be.db.respository;
+package trendravel.photoravel_be.db.respository.spot;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import trendravel.photoravel_be.db.spot.Spot;
